requirements.txt
astor==0.8.1
astroid==2.15.1
attrs==22.2.0
bandit==1.7.5
certifi==2022.12.7
click==8.1.3
darglint==1.8.1
dill==0.3.6
docutils==0.19
eradicate==2.2.0
flake8==4.0.1
flake8-bandit==3.0.0
flake8-broken-line==0.5.0
flake8-bugbear==22.12.6
flake8-commas==2.1.0
flake8-comprehensions==3.11.1
flake8-debugger==4.1.2
flake8-docstrings==1.7.0
flake8-eradicate==1.4.0
flake8-html==0.4.3
flake8-isort==4.2.0
flake8-polyfill==1.0.2
flake8-quotes==3.3.2
flake8-rst-docstrings==0.2.7
flake8-string-format==0.3.0
Flask==2.2.3
gevent==22.10.2
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
importlib-metadata==6.1.0
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mccabe==0.6.1
mdurl==0.1.2
mypy==1.1.1
mypy-extensions==1.0.0
packaging==23.0
pbr==5.11.1
pep8-naming==0.13.2
platformdirs==3.2.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pydantic==1.10.7
pydocstyle==6.3.0
pyflakes==2.4.0
Pygments==2.14.0
pytest==7.0.0
pytest-asyncio==0.20.1
pytest-metadata==2.0.4
python-dotenv==1.0.0
python-logstash==0.4.8
PyYAML==6.0
restructuredtext-lint==1.4.0
rich==13.3.3
sentry-sdk==1.19.1
smmap==5.0.0
snowballstemmer==2.2.0
spectree==1.1.0
stevedore==5.0.0
tomli==2.0.1
tomlkit==0.11.7
typing_extensions==4.5.0
urllib3==1.26.15
wemake-python-styleguide==0.17.0
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
gunicorn==20.1.0
sentry-sdk[flask]
pymongo
clickhouse-driver==0.2.5



